Added an alternative hand evaluator that is roughly 3 to 4 times faster than our current one. Fixed tests including Cards. Added a new project called hand-eval.

+++ trunk/server/common/src/test/java/org/cspoker/server/common/game/elements/cards/hand/TestHandQuality.java
@@ -241,8 +241,8 @@ public class TestHandQuality extends TestCase {
 			qualityGreater = (compare == 1);
 			qualitySmaller = (compare == -1);
 			qualityEqual = (compare == 0);
-			quality1 = hand1.getRank().intValue();
-			quality2 = hand2.getRank().intValue();
+			quality1 = hand1.getRank();
+			quality2 = hand2.getRank();
 
 			String rank = hand1.getShortDescription();
 			if (!rankMap.containsKey(rank)) {
